dodalam timestamp do histogramu i mam dystans

+++ app/src/main/java/pl/roszkowska/track/database/StepEntity.java
@@ -10,6 +10,6 @@ public class StepEntity {
     long routeId;
     double lat;
     double lon;
-    double timestamp;
+    long timestamp;
     double distanceBetweenLastStep;
 }
